learn laravel part 6

+++ app/Http/Controllers/JobController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\JobPosted;
 use App\Models\Job;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -37,11 +39,16 @@ class JobController extends Controller
             'salary' => ['required']
         ]);
 
-        Job::create([
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
+
+        Mail::to($job->employer->user)->send(
+            new JobPosted($job)
+        );
+
 
         return redirect('/jobs');
     }
@@ -64,7 +71,7 @@ class JobController extends Controller
         // }
 
         // lanjutan method 2
-        Gate::authorize('edit-job', $job);
+        // Gate::authorize('edit-job', $job); // auth on controller level
 
         return view('jobs.edit', ['job' => $job]);
     }
@@ -77,6 +84,8 @@ class JobController extends Controller
         ]);
 
         // authorize
+        Gate::authorize('edit-job', $job);
+
         // update the job and persist
         $job->update([
             'title' => request('title'),
@@ -90,6 +99,7 @@ class JobController extends Controller
     public function destroy(Job $job)
     {
         // authorize
+        Gate::authorize('edit-job', $job);
 
         // delete the job
         $job->delete();
